Motors are like butter

+++ motor_thread.c
@@ -51,7 +51,7 @@
         if (val > 100) val = 100;
         motor.leftPower = val*DRV_TMR0_PeriodValueGet()/100;
         motor.leftDir = dir;
-        PLIB_OC_PulseWidth16BitSet(OC_ID_1, val*DRV_TMR0_PeriodValueGet()/100);
+        PLIB_OC_PulseWidth16BitSet(OC_ID_1, (uint16_t)val*DRV_TMR0_PeriodValueGet()/100);
         PLIB_PORTS_PinWrite (PORTS_ID_0, PORT_CHANNEL_C, PORTS_BIT_POS_14, !dir);
         DRV_OC0_Start();
         return;
@@ -62,7 +62,7 @@
         if (val > 100) val = 100;
         motor.rightPower = val*DRV_TMR0_PeriodValueGet()/100;
         motor.rightDir = dir;
-        PLIB_OC_PulseWidth16BitSet(OC_ID_2, val*DRV_TMR0_PeriodValueGet()/100);
+        PLIB_OC_PulseWidth16BitSet(OC_ID_2, (uint16_t)val*DRV_TMR0_PeriodValueGet()/100);
         PLIB_PORTS_PinWrite (PORTS_ID_0, PORT_CHANNEL_G, PORTS_BIT_POS_1, !dir);
         DRV_OC1_Start();
         return;
